Work on remarks
